VS compilation.

+++ rEFIt_UEFI/Platform/Posix/strlen.cpp
@@ -4,6 +4,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include <stddef.h>
+#include <limits.h>
 
 #ifdef MAP_TO_EDK_STRLEN
 extern "C" UINTN EFIAPI AsciiStrLen (IN CONST CHAR8 *String);
@@ -23,7 +24,7 @@ size_t strlen (const char *str)
 
   /* Handle the first few characters by reading one character at a time.
      Do this until CHAR_PTR is aligned on a longword boundary.  */
-  for (char_ptr = str; ((unsigned long int) char_ptr
+  for (char_ptr = str; ((uintptr_t) char_ptr
                         & (sizeof (longword) - 1)) != 0;
        ++char_ptr)
     if (*char_ptr == '\0') {
